[chore] add sendTokenSteps page
