+++ src/pages/SendTokenStepsPage/components/FailComponent.js
