+++ src/pages/SendTokenStepsPage/components/FinalConfirmation.js
