+++ src/pages/SendTokenStepsPage/components/PlatformList.js
