+++ src/pages/SendTokenStepsPage/components/SendSuccess.js
