+++ src/pages/SendTokenStepsPage/components/Step1.js
