+++ src/pages/SendTokenStepsPage/components/Step2.js
