+++ src/pages/SendTokenStepsPage/components/TokenBottomModal.js
