+++ src/pages/SendTokenStepsPage/components/connection/index.js
