+++ src/pages/SendTokenStepsPage/components/index.js
@@ -1,0 +1,11 @@
+export { default as LoadingComponent } from "components/card/LoadingComponent";
+export { default as CheckSendModal } from "./CheckSendModal";
+export { default as FailedComponent } from "./FailComponent";
+export { default as FinalConfirmation } from "./FinalConfirmation";
+export { default as MobileNetworkBox } from "./MobileNetworkBox";
+export { default as NetworkSwitchModal } from "./NetworkSwitchModal";
+export { default as SendSuccess } from "./SendSuccess";
+export { default as Step1 } from "./Step1";
+export { default as Step2 } from "./Step2";
+export { default as Step3 } from "./Step3";
+export { default as TokenBottomModal } from "./TokenBottomModal";
