+++ src/pages/SendTokenStepsPage/data/SimpleTokenList.js
