+++ src/pages/SendTokenStepsPage/data/SimpleTokenList2.js
