+++ src/pages/SettingPage/components/EnrolledAccount.js
@@ -1,4 +1,4 @@
-import PlatformList from "pages/SendTokenPage/components/PlatformList";
+import PlatformList from "pages/SendTokenStepsPage/components/PlatformList";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
